Relocate root test-config classes into util
Move SpringBootConfigurationTests (renamed JsonFormatMapperTestConfig),
TestMessagingConfig, and TestJmsTemplateMockConfig alongside SpringBootTestContext;
drop the now same-package BaseMessagingIntTest import.
