+++ src/test/java/com/otilm/core/util/JsonFormatMapperTestConfig.java
@@ -1,4 +1,4 @@
-package com.otilm.core;
+package com.otilm.core.util;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.hibernate.cfg.AvailableSettings;
@@ -10,7 +10,7 @@ import org.springframework.context.annotation.Profile;
 
 @Configuration
 @Profile("test")
-public class SpringBootConfigurationTests {
+public class JsonFormatMapperTestConfig {
 
     @Bean
     HibernatePropertiesCustomizer jsonFormatMapperCustomizer(ObjectMapper objectMapper) {
